RFC: Add CODEOWNERS for v1
This PR adds a CODEOWNERS to be used against the v1 branch for its lifetime until merged with main.

It is **NOT** intended to be used to enforce or mandate blocking individuals for merge. Listings can / **SHOULD** be modified based on interest / desire for code reviews, and **SHOULD NOT** be considered exhaustive.